pyproject.toml
[project]
name = "unet-web-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.2",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pathlib>=1.0.1",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
]
